requirements.txt
alembic==1.7.5
alipay-sdk-python==3.3.398
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.15.0
amqp==5.0.9
aniso8601==9.0.1
apispec==5.1.1
apispec-webframeworks==0.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
billiard==3.6.4.0
black==21.12b0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crcmod==1.7
cryptography==36.0.1
decorator==5.1.1
Deprecated==1.2.13
executing==0.8.2
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-RESTy==2.0.0
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
ipython==8.0.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonmerge==1.8.0
jsonschema==4.4.0
kombu==5.2.3
konch==4.4.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
matplotlib-inline==0.1.3
mistune==2.0.2
mypy-extensions==0.4.3
numpy==1.22.1
optionaldict==0.1.2
oss2==2.15.0
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.14.0
pycryptodomex==3.9.4
Pygments==2.11.2
pyOpenSSL==19.1.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.1.1
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
stack-data==0.1.4
tomli==1.2.3
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wechatpy==1.8.18
Werkzeug==2.0.2
wrapt==1.13.3
xlwt==1.3.0
xmltodict==0.12.0
zope.event==4.5.0
zope.interface==5.4.0
